refactor(backend): move $defs inliner to api_utils
